src/requirements.txt
flask
pytest
request